requirements.txt
#Core Libraries
langchain
llama-index
haystack

#for a backend server
fastapi
uvicorn

#Additional Tools
requests
python-dotenv

#optional Ai agent framework(intall if want to usethem)
autogpt
babyagi